# perf: PeriodCombiner improvements