--- a/lib/dash/dash_parser.js
+++ b/lib/dash/dash_parser.js
@@ -1610,7 +1610,6 @@ shaka.dash.DashParser = class {
         hdr,
         videoLayout: undefined,
         tilesLayout,
-        matchedStreams: [],
         accessibilityPurpose,
         external: false,
         fastSwitching: false,