--- a/lib/util/periods.js
+++ b/lib/util/periods.js
@@ -826,6 +826,8 @@ shaka.util.PeriodCombiner = class {
       }
     };
 
+    // Clone roles array so this output stream can own it.
+    clone.roles = clone.roles.slice();
     clone.segmentIndex = null;
     clone.emsgSchemeIdUris = [];
     clone.keyIds = new Set();
@@ -871,7 +873,17 @@ shaka.util.PeriodCombiner = class {
     // the line.
 
     // Combine arrays, keeping only the unique elements
-    const combineArrays = (a, b) => Array.from(new Set(a.concat(b)));
+    const combineArrays = (output, input) => {
+      if (!output) {
+        output = [];
+      }
+      for (const item of input) {
+        if (!output.includes(item)) {
+          output.push(item);
+        }
+      }
+      return output;
+    };
     output.roles = combineArrays(output.roles, input.roles);
 
     if (input.emsgSchemeIdUris) {
@@ -879,8 +891,9 @@ shaka.util.PeriodCombiner = class {
           output.emsgSchemeIdUris, input.emsgSchemeIdUris);
     }
 
-    const combineSets = (a, b) => new Set([...a, ...b]);
-    output.keyIds = combineSets(output.keyIds, input.keyIds);
+    for (const keyId of input.keyIds) {
+      output.keyIds.add(keyId);
+    }
 
     if (output.originalId == null) {
       output.originalId = input.originalId;
@@ -952,11 +965,22 @@ shaka.util.PeriodCombiner = class {
    */
   static concatenateStreamDBs_(output, input) {
     // Combine arrays, keeping only the unique elements
-    const combineArrays = (a, b) => Array.from(new Set(a.concat(b)));
+    const combineArrays = (output, input) => {
+      if (!output) {
+        output = [];
+      }
+      for (const item of input) {
+        if (!output.includes(item)) {
+          output.push(item);
+        }
+      }
+      return output;
+    };
     output.roles = combineArrays(output.roles, input.roles);
 
-    const combineSets = (a, b) => new Set([...a, ...b]);
-    output.keyIds = combineSets(output.keyIds, input.keyIds);
+    for (const keyId of input.keyIds) {
+      output.keyIds.add(keyId);
+    }
 
     // The output is encrypted if any input was encrypted.
     output.encrypted = output.encrypted && input.encrypted;