--- a/lib/util/periods.js
+++ b/lib/util/periods.js
@@ -506,8 +506,6 @@ shaka.util.PeriodCombiner = class {
    */
   async combine_(
       outputStreams, streamsPerPeriod, firstNewPeriodIndex, clone, concat) {
-    const ContentType = shaka.util.ManifestParserUtils.ContentType;
-
     const unusedStreamsPerPeriod = [];
     for (let i = 0; i < streamsPerPeriod.length; i++) {
       if (i >= firstNewPeriodIndex) {
@@ -556,10 +554,7 @@ shaka.util.PeriodCombiner = class {
 
     for (const unusedStreams of unusedStreamsPerPeriod) {
       for (const stream of unusedStreams) {
-        const isDummyText = stream.type == ContentType.TEXT && !stream.language;
-        const isDummyImage = stream.type == ContentType.IMAGE &&
-            !stream.tilesLayout;
-        if (isDummyText || isDummyImage) {
+        if (shaka.util.PeriodCombiner.isDummy_(stream)) {
           // This is one of our dummy streams, so ignore it.  We may not use
           // them all, and that's fine.
           continue;
@@ -707,6 +702,11 @@ shaka.util.PeriodCombiner = class {
    */
   createNewOutputStream_(
       stream, streamsPerPeriod, clone, concat, unusedStreamsPerPeriod) {
+    // Check do we want to create output stream from dummy stream
+    // and if so, return quickly.
+    if (shaka.util.PeriodCombiner.isDummy_(stream)) {
+      return null;
+    }
     // Start by cloning the stream without segments, key IDs, etc.
     const outputStream = clone(stream);
 
@@ -1152,14 +1152,6 @@ shaka.util.PeriodCombiner = class {
     // For text, we don't care about MIME type or codec.  We can always switch
     // between text types.
 
-    // The output stream should not be a dummy stream inserted to fill a period
-    // gap.  So reject any candidate if the output has no language.  This would
-    // cause findMatchesInAllPeriods_ to return null and this output stream to
-    // be skipped (meaning no output streams based on it).
-    if (!outputStream.language) {
-      return false;
-    }
-
     // If the candidate is a dummy, then it is compatible, and we could use it
     // if nothing else matches.
     if (!candidate.language) {
@@ -1193,14 +1185,6 @@ shaka.util.PeriodCombiner = class {
   static areImageStreamsCompatible_(outputStream, candidate) {
     // For image, we don't care about MIME type.  We can always switch
     // between image types.
-
-    // The output stream should not be a dummy stream inserted to fill a period
-    // gap.  So reject any candidate if the output has no tilesLayout.  This
-    // would cause findMatchesInAllPeriods_ to return null and this output
-    // stream to be skipped (meaning no output streams based on it).
-    if (!outputStream.tilesLayout) {
-      return false;
-    }
 
     return true;
   }
@@ -1708,6 +1692,25 @@ shaka.util.PeriodCombiner = class {
     }
 
     return EQUAL;
+  }
+
+  /**
+   * @param {T} stream
+   * @return {boolean}
+   * @template T
+   * Accepts either a StreamDB or Stream type.
+   * @private
+   */
+  static isDummy_(stream) {
+    const ContentType = shaka.util.ManifestParserUtils.ContentType;
+    switch (stream.type) {
+      case ContentType.TEXT:
+        return !stream.language;
+      case ContentType.IMAGE:
+        return !stream.tilesLayout;
+      default:
+        return false;
+    }
   }
 };