--- a/lib/util/periods.js
+++ b/lib/util/periods.js
@@ -507,14 +507,13 @@ shaka.util.PeriodCombiner = class {
   async combine_(
       outputStreams, streamsPerPeriod, firstNewPeriodIndex, clone, concat) {
     const unusedStreamsPerPeriod = [];
-    for (let i = 0; i < streamsPerPeriod.length; i++) {
-      if (i >= firstNewPeriodIndex) {
-        // This periods streams are all new.
-        unusedStreamsPerPeriod.push(new Set(streamsPerPeriod[i]));
-      } else {
-        // This period's streams have all been used already.
-        unusedStreamsPerPeriod.push(new Set());
-      }
+    for (let i = 0; i < firstNewPeriodIndex; i++) {
+      // This period's streams have all been used already.
+      unusedStreamsPerPeriod.push(new Set());
+    }
+    for (let i = firstNewPeriodIndex; i < streamsPerPeriod.length; i++) {
+      // This periods streams are all new.
+      unusedStreamsPerPeriod.push(new Set(streamsPerPeriod[i]));
     }
 
     // First, extend all existing output Streams into the new periods.