--- a/lib/util/periods.js
+++ b/lib/util/periods.js
@@ -118,6 +118,8 @@ shaka.util.PeriodCombiner = class {
    * @private
   */
   getStreamsPerPeriod_(periods) {
+    const ContentType = shaka.util.ManifestParserUtils.ContentType;
+
     const audioStreamsPerPeriod = [];
     const videoStreamsPerPeriod = [];
     const textStreamsPerPeriod = [];
@@ -128,6 +130,16 @@ shaka.util.PeriodCombiner = class {
       videoStreamsPerPeriod.push(period.videoStreams);
       textStreamsPerPeriod.push(period.textStreams);
       imageStreamsPerPeriod.push(period.imageStreams);
+
+      // It's okay to have a period with no text or images, but our algorithm
+      // fails on any period without matching streams.  So we add dummy streams
+      // to each period.  Since we combine text streams by language and image
+      // streams by resolution, we might need a dummy even in periods with these
+      // streams already.
+      period.textStreams.push(shaka.util.PeriodCombiner.dummyStream_(
+          ContentType.TEXT));
+      period.imageStreams.push(shaka.util.PeriodCombiner.dummyStream_(
+          ContentType.IMAGE));
     }
     return {
       audioStreamsPerPeriod,
@@ -188,20 +200,6 @@ shaka.util.PeriodCombiner = class {
         textStreamsPerPeriod,
         imageStreamsPerPeriod,
       } = this.getStreamsPerPeriod_(periods);
-
-      // It's okay to have a period with no text or images, but our algorithm
-      // fails on any period without matching streams.  So we add dummy streams
-      // to each period.  Since we combine text streams by language and image
-      // streams by resolution, we might need a dummy even in periods with these
-      // streams already.
-      for (const textStreams of textStreamsPerPeriod) {
-        textStreams.push(shaka.util.PeriodCombiner.dummyStream_(
-            ContentType.TEXT));
-      }
-      for (const imageStreams of imageStreamsPerPeriod) {
-        imageStreams.push(shaka.util.PeriodCombiner.dummyStream_(
-            ContentType.IMAGE));
-      }
 
       await Promise.all([
         this.combine_(