--- a/lib/util/periods.js
+++ b/lib/util/periods.js
@@ -785,6 +785,8 @@ shaka.util.PeriodCombiner = class {
     const clone = /** @type {shaka.extern.StreamDB} */(Object.assign(
         {}, streamDb));
 
+    // Clone roles array so this output stream can own it.
+    clone.roles = clone.roles.slice();
     // These are wiped out now and rebuilt later from the various per-period
     // streams that match this output.
     clone.keyIds = new Set();